Split validate.js helpers into smaller modules
